Split up parsing tests
